create dbconfig.py
