+++ static/dialog_script.js
@@ -158,7 +158,7 @@ async function signon() {
     };
 
     try {
-        let response = await fetch('http://127.0.0.1:8000/api/user', {
+        let response = await fetch('http://54.79.121.157:8000/api/user', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
